@@ -3,10 +3,10 @@
   "version": "1.0.0",
   "description": "JavaScript development environment Pluralsight course by Cory House",
   "scripts": {
-    "prestart": "node buildScripts/startMessage",
+    "prestart": "babel-node buildScripts/startMessage.js",
     "start": "npm-run-all --parallel open:src",
     "share": "lt --port 3000",
-    "open:src": "node buildScripts/srcServer.js"
+    "open:src": "babel-node buildScripts/srcServer.js"
   },
   "author": "Cory House",
   "license": "MIT",
@@ -14,36 +14,39 @@
     "whatwg-fetch": "1.0.0"
   },
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
+    "@babel/cli": "^7.0.0-rc.1",
+    "@babel/core": "^7.0.0-rc.1",
+    "@babel/node": "^7.0.0-rc.1",
+    "@babel/preset-env": "^7.0.0-rc.1",
     "babel-loader": "6.2.5",
     "babel-preset-latest": "6.16.0",
     "babel-register": "6.16.3",
     "chai": "3.5.0",
     "chalk": "1.1.3",
     "cheerio": "0.22.0",
-    "compression": "1.6.2",
+    "compression": "^1.7.4",
     "cross-env": "3.1.3",
-    "css-loader": "0.25.0",
+    "css-loader": "^2.1.1",
     "eslint": "3.8.1",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-watch": "2.1.14",
-    "express": "4.14.0",
+    "eslint-plugin-import": "^2.17.2",
+    "eslint-watch": "^5.1.2",
+    "express": "^4.16.4",
     "extract-text-webpack-plugin": "1.0.1",
-    "html-webpack-plugin": "2.22.0",
+    "html-webpack-plugin": "^3.2.0",
     "jsdom": "9.8.0",
-    "json-schema-faker": "0.3.6",
-    "json-server": "0.8.22",
-    "localtunnel": "1.8.1",
-    "mocha": "3.1.2",
-    "nock": "8.1.0",
+    "json-schema-faker": "^0.4.7",
+    "json-server": "^0.14.2",
+    "localtunnel": "^1.9.1",
+    "mocha": "^6.1.4",
+    "nock": "^10.0.6",
+    "nodemon": "^1.18.3",
     "npm-run-all": "3.1.1",
-    "nsp": "2.6.2",
+    "nsp": "^3.2.1",
     "numeral": "1.5.3",
-    "open": "0.0.5",
+    "open": "^6.2.0",
     "rimraf": "2.5.4",
     "style-loader": "0.13.1",
-    "webpack": "1.13.2",
+    "webpack": "^4.30.0",
     "webpack-dev-middleware": "1.8.4",
     "webpack-hot-middleware": "2.13.0",
     "webpack-md5-hash": "0.0.5"
