@@ -1,3 +1,3 @@
-var chalk = require('chalk');
+import chalk from 'chalk';
 
 console.log(chalk.green('Starting app in dev mode...'));
